pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "homedeck"
version = "0.1.0"
description = "Control Home Assistant using Stream Deck-like devices"
authors = [{name = "redphx"}]
readme = {file = "README.md", content-type = "text/markdown"}
license = {file = "LICENSE"}
keywords = ["streamdeck", "stream deck", "stream-deck"]
requires-python = ">=3.9"
dependencies = [
    "strmdck==0.1.0",
    "cairosvg==2.7.1",
    "construct==2.10.70",
    "deepdiff==8.2.0",
    "fastapi[standard]==0.115.11",
    "httpx==0.28.1",
    "Jinja2==3.1.5",
    "jsonschema==4.23.0",
    "pillow==11.1.0",
    "psutil==7.0.0",
    "pytest==8.3.4",
    "python-dotenv==1.0.1",
    "PyYAML==6.0.2",
    "toml==0.10.2",
    "uvicorn==0.34.0",
    "watchdog==6.0.0",
    "websockets==14.2",
    "zeroconf==0.146.1"
]

[project.urls]
Homepage = "https://github.com/redphx/strmdck-home-assistant"
Issues = "https://github.com/redphx/strmdck-home-assistant/issues"
Repository = "https://github.com/redphx/strmdck-home-assistant.git"
